+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQ70ETS860YYW972QPWM2AHJ"
+        "01KQ7Q5D4HNF05TW91GQJF5XMH"
     )
     r"""The unique identifier of the data source"""