+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQ70ETS6CJZ82F4C0QV6YH6D"
+        "01KQ7Q5D4F5DV21S7RW2PVNH03"
     )
     r"""The unique identifier of the data source"""