+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQ70ETSFCQ5CKSSZS16JJJNZ"
+        "01KQ7Q5D4KDP5FEWY09K4900PZ"
     )
     r"""The unique identifier of the data source"""