+++ packages/orq-rc/RELEASES.md
@@ -7349,3 +7349,13 @@ Based on:
 - [python v4.9.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.2 - packages/orq-rc
+
+## 2026-04-27 15:01:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.3 - packages/orq-rc